+++ ciao-controller/internal/datastore/datastore.go
@@ -272,6 +272,15 @@ func (ds *Datastore) Init(config Config) error {
 			ds.nodes[i.NodeID] = n
 		}
 		ds.nodes[i.NodeID].instances[key] = i
+
+		// ds.tenants.instances should point to the same
+		// instances that we have in ds.instances, otherwise they
+		// will not get updated when we get new stats.
+
+		tenant := ds.tenants[i.TenantID]
+		if tenant != nil {
+			tenant.instances[i.ID] = i
+		}
 	}
 
 	ds.tenantUsage = make(map[string][]types.CiaoUsage)
@@ -298,6 +307,11 @@ func (ds *Datastore) Init(config Config) error {
 		}
 
 		ds.instanceVolumes[link] = key
+
+		instance := ds.instances[value.InstanceID]
+		if instance != nil {
+			instance.Attachments = append(instance.Attachments, value)
+		}
 	}
 
 	ds.attachLock = &sync.RWMutex{}
@@ -907,7 +921,12 @@ func (ds *Datastore) StopFailure(instanceID string, reason payloads.StopFailureR
 // for this tenant. If the instance was a normal tenant instance, the
 // IP address will be released and the instance will be deleted from the
 // datastore.
-func (ds *Datastore) StartFailure(instanceID string, reason payloads.StartFailureReason) error {
+//
+// Only instances whose status is pending are removed when a StartFailure event
+// is received.  StartFailure errors may also be generated when restarting an
+// exited instance and we want to make sure that a failure to restart such
+// an instance does not result in it being deleted.
+func (ds *Datastore) StartFailure(instanceID string, reason payloads.StartFailureReason, migration bool) error {
 	var tenantID string
 	var cnci bool
 
@@ -952,7 +971,7 @@ func (ds *Datastore) StartFailure(instanceID string, reason payloads.StartFailur
 		return errors.Wrapf(err, "error getting instance (%v)", instanceID)
 	}
 
-	if reason.IsFatal() {
+	if reason.IsFatal() && !migration {
 		ds.deleteInstance(instanceID)
 	}
 
@@ -1035,7 +1054,9 @@ func (ds *Datastore) deleteInstance(instanceID string) (string, error) {
 
 	ds.tenantsLock.Lock()
 	tenant := ds.tenants[i.TenantID]
-	delete(tenant.instances, instanceID)
+	if tenant != nil {
+		delete(tenant.instances, instanceID)
+	}
 	ds.tenantsLock.Unlock()
 
 	// we may not have received any node stats for this instance
@@ -1073,6 +1094,83 @@ func (ds *Datastore) DeleteInstance(instanceID string) error {
 	msg := fmt.Sprintf("Deleted Instance %s", instanceID)
 	ds.db.logEvent(tenantID, string(userInfo), msg)
 
+	return nil
+}
+
+func (ds *Datastore) restartInstance(instanceID string) error {
+	ds.instancesLock.Lock()
+	i := ds.instances[instanceID]
+	i.State = payloads.Pending
+	ds.instancesLock.Unlock()
+
+	instanceStat := types.CiaoServerStats{
+		ID:        instanceID,
+		Timestamp: time.Now(),
+		Status:    payloads.Pending,
+	}
+	ds.instanceLastStatLock.Lock()
+	ds.instanceLastStat[instanceID] = instanceStat
+	ds.instanceLastStatLock.Unlock()
+
+	stats := []payloads.InstanceStat{
+		{
+			InstanceUUID: instanceID,
+			State:        payloads.Pending,
+		},
+	}
+
+	return errors.Wrapf(ds.db.addInstanceStats(stats, ""), "error adding instance stats to database")
+}
+
+// RestartInstance resets a restarting instance's state to pending.
+func (ds *Datastore) RestartInstance(instanceID string) error {
+	err := ds.restartInstance(instanceID)
+	if err != nil {
+		return errors.Wrapf(err, "error restarting instance")
+	}
+	return nil
+}
+
+func (ds *Datastore) stopInstance(instanceID string) error {
+	ds.instancesLock.Lock()
+	i := ds.instances[instanceID]
+	oldNodeID := i.NodeID
+	i.NodeID = ""
+	i.State = payloads.Exited
+	ds.instancesLock.Unlock()
+
+	// we may not have received any node stats for this instance
+	if oldNodeID != "" {
+		ds.nodesLock.Lock()
+		delete(ds.nodes[oldNodeID].instances, instanceID)
+		ds.nodesLock.Unlock()
+	}
+
+	instanceStat := types.CiaoServerStats{
+		ID:        instanceID,
+		Timestamp: time.Now(),
+		Status:    payloads.Exited,
+	}
+	ds.instanceLastStatLock.Lock()
+	ds.instanceLastStat[instanceID] = instanceStat
+	ds.instanceLastStatLock.Unlock()
+
+	stats := []payloads.InstanceStat{
+		{
+			InstanceUUID: instanceID,
+			State:        payloads.Exited,
+		},
+	}
+
+	return errors.Wrapf(ds.db.addInstanceStats(stats, ""), "error adding instance stats to database")
+}
+
+// StopInstance removes the link between an instance and its node
+func (ds *Datastore) StopInstance(instanceID string) error {
+	err := ds.stopInstance(instanceID)
+	if err != nil {
+		return errors.Wrapf(err, "error stopping instance")
+	}
 	return nil
 }
